get(), post(), put() and delete() append to path
get(), post(), put() and delete() append to path rather than replacing it, so that if path is set in self::create, it is used as the base.

e.g. if /api/v1 is specified as part of the url in create, then you use get('/clients'), it results in an endpoint of /api/v1/clients.